Revert "Refactor/command buffer"
