+++ SSD_Project/FileManager.h
@@ -20,24 +20,30 @@ public:
     virtual void readNand(unsigned int line);
     virtual void writeResult(string value);
 
+    virtual void writeBuffer(string command);
+    virtual void removeBuffer(string command);
+    virtual vector<string> readBuffer();
+    virtual void flushBuffer();
 private:
 #ifndef __TEST__
     FileManager();
 #endif
     FileManager& operator=(const FileManager& otherInstance) = delete;
     FileManager(const FileManager& otherInstance) = delete;
+    void createFile(string fileName);
     ifstream getNandFile();
     ofstream getResultFile();
+    ifstream getCommandBuffer();
     bool checkValidLineNum(int line);
     bool getAllData(ifstream& readFile, vector<string>& readLines);
     void setAllData(ofstream& writeFile, vector<string>& readLines);
-    void createFile(string fileName);
-
     string getData(ifstream& nandFile, unsigned int line);
 
     const int MAX_LINE_NUM = 99;
     const int MIN_LINE_NUM = 0;
+    const int MAX_COMMAND_NUM_IN_BUFFER = 10;
     const string NAND_FILE_NAME = "nand.txt";
     const string RESULT_FILE_NAME = "result.txt";
+    const string COMMAND_BUFFER_NAME = "buffer.txt";
     const string DEFAULT_WRITE_VALUE = "0x00000000";
 };
