+++ SSD_Project/SSD_Project.vcxproj
@@ -131,14 +131,12 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="BufferManager.cpp" />
     <ClCompile Include="CommandManager.cpp" />
     <ClCompile Include="FileManager.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="Ssd.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="BufferManager.h" />
     <ClInclude Include="CommandManager.h" />
     <ClInclude Include="SsdDriver.h" />
     <ClInclude Include="util.h" />
