+++ SSD_Project/Ssd.cpp
@@ -1,7 +1,7 @@
 #pragma once
 #include "SsdDriver.h"
 #include "FileManager.h"
-#include "BufferManager.h"
+#include "util.h"
 #include "../Logger/Logger.h"
 
 #include <iostream>
@@ -16,45 +16,52 @@ class Ssd : public SsdDriver {
 public:
     Ssd() {
         fileManager = &FileManager::getInstance();
-        bufferManager = &BufferManager::getInstance();
+    }
+
+    void setFileManager(FileManager* newFileManager) {
+        fileManager = newFileManager;
+    }
+
+    string writeCommand(unsigned int line, string value) {
+        return "W " + to_string(line) + " " + value;
+    }
+
+    string eraseCommand(unsigned int line, unsigned int size) {
+        return "E " + to_string(line) + " " + to_string(size);
     }
 
     void write(unsigned int line, string value) override {
-        string writeCmdString = bufferManager->writeCommand(line, value);
-        vector<string> cmdStrings = bufferManager->readBuffer();
-        vector<string> eraseCmds;
+        vector<string> cmdStrings = fileManager->readBuffer();
+        vector<Command> eraseCmds;
         for (auto cmdString : cmdStrings) {
-            char cmdType = bufferManager->getCmdType(cmdString);
-            int lba = bufferManager->getCmdLba(cmdString); 
-            if (cmdString == writeCmdString)
+            if (cmdString == writeCommand(line, value))
                 return;
-            if (cmdType == 'W' && lba == line) {
-                bufferManager->removeBuffer(cmdString);
+            Command cmd = cmdFormat.parseCommand(cmdString);
+            if (cmd.type == 'W' && cmd.lba == line) {
+                fileManager->removeBuffer(writeCommand(cmd.lba, cmd.value));
             }
-            else if (cmdType == 'E') {
-                eraseCmds.push_back(cmdString);
+            else if (cmd.type == 'E') {
+                eraseCmds.push_back(cmd);
             }
         }
         if (cmdStrings.size() >= MAX_COMMAND_NUM_IN_BUFFER)
             flush();
-        bufferManager->writeBuffer(writeCmdString);
-        bufferManager->updateErase(eraseCmds);
+        fileManager->writeBuffer(writeCommand(line, value));
+        updateErase(eraseCmds);
     }
 
     void read(unsigned int line) override {
-        vector<string> cmdStrings = bufferManager->readBuffer();
+        vector<string> cmdStrings = fileManager->readBuffer();
         for (auto cmdString : cmdStrings) {
-            char cmdType = bufferManager->getCmdType(cmdString);
-            int lba = bufferManager->getCmdLba(cmdString);
-            if (cmdType == 'W' && lba== line) {
-                string value = bufferManager->getCmdValue(cmdString);
-                fileManager->writeResult(value);
-                PRINT_LOG(("Buffer Hit Read succeeded / LBA: " + to_string(line) + " / Value: " + value));
+            Command cmd = cmdFormat.parseCommand(cmdString);
+
+            if (cmd.type == 'W' && cmd.lba == line) {
+                fileManager->writeResult(cmd.value);
+                PRINT_LOG(("Buffer Hit Read succeeded / LBA: " + to_string(line) + " / Value: " + cmd.value));
                 return;
             }
-            else if (cmdType == 'E') {
-                int size = bufferManager->getCmdSize(cmdString);
-                if (line >= lba && line < lba + size) {
+            else if (cmd.type == 'E') {
+                if (line >= cmd.lba && line < cmd.lba + cmd.size) {
                     fileManager->writeResult(DEFAULT_VALUE);
                     return;
                 }
@@ -63,63 +70,114 @@ public:
         fileManager->readNand(line);
     }
 
+    static bool compareByLba(const Command& a, const Command& b) {
+        return a.lba < b.lba;
+    }
+
     void erase(unsigned int line, unsigned int size) override {
         size = min(MAX_ERASE_SIZE, size);
-        vector<string> cmdStrings = bufferManager->readBuffer();
-        vector<string> eraseCmds;
+        vector<string> cmdStrings = fileManager->readBuffer();
+        vector<Command> eraseCmds;
         for (auto cmdString : cmdStrings) {
-            if (bufferManager->getCmdType(cmdString) == 'W') {
-                if (bufferManager->getCmdLba(cmdString) >= line
-                    && bufferManager->getCmdLba(cmdString)< line + size) {
-                    bufferManager->removeBuffer(cmdString);
+            Command cmd = cmdFormat.parseCommand(cmdString);
+            if (cmd.type == 'W') {
+                if (cmd.lba >= line && cmd.lba < line + size) {
+                    fileManager->removeBuffer(writeCommand(cmd.lba, cmd.value));
                 }
             }
-            else if (bufferManager->getCmdType(cmdString) == 'E') {
-                if (bufferManager->getCmdLba(cmdString) == line 
-                    && bufferManager->getCmdSize(cmdString) == size)
+            else if (cmd.type == 'E') {
+                if (cmd.lba == line && cmd.size == size)
                     return;
-                eraseCmds.push_back(cmdString);
+                eraseCmds.push_back(cmd);
             }
         }
-        if (bufferManager->readBuffer().size() >= MAX_COMMAND_NUM_IN_BUFFER)
+        if (fileManager->readBuffer().size() >= MAX_COMMAND_NUM_IN_BUFFER)
             flush();
-        string eraseCmdString = bufferManager->eraseCommand(line, size);
-        bufferManager->writeBuffer(eraseCmdString);
 
-        eraseCmds.push_back(eraseCmdString);
-        bufferManager->mergeConsecutiveErases(eraseCmds);
+        fileManager->writeBuffer(eraseCommand(line, size));
+
+        eraseCmds.push_back(cmdFormat.parseCommand(eraseCommand(line, size)));
+        mergeConsecutiveErases(eraseCmds);
     }
+
+    void mergeConsecutiveErases(std::vector<Command>& eraseCmds) {
+        std::sort(eraseCmds.begin(), eraseCmds.end(), compareByLba);
+        int startLine = 0, endLine = 0, eraseSize = 0;
+        bool consecutive = false;
+        for (int i = 0; i < eraseCmds.size(); i++) {
+            if (consecutive == false) {
+                startLine = eraseCmds[i].lba;
+                eraseSize = eraseCmds[i].size;
+                endLine = eraseCmds[i].lba + eraseSize;
+                consecutive = true;
+                continue;
+            }
+            if (eraseCmds[i].lba == endLine) {
+                fileManager->removeBuffer(eraseCommand(startLine, eraseSize));
+                fileManager->removeBuffer(eraseCommand(eraseCmds[i].lba, eraseCmds[i].size));
+                int newSize = eraseCmds[i].size + eraseSize;
+                fileManager->writeBuffer(eraseCommand(startLine, newSize));
+                eraseSize = newSize;
+                endLine = startLine + eraseSize;
+                consecutive = true;
+            }
+            else {
+                consecutive = false;
+            }
+        }
+    }
+
+    void updateErase(std::vector<Command>& eraseCmds) {
+        vector<string> cmdStrings = fileManager->readBuffer();
+        for (auto cmdString : cmdStrings) {
+            Command cmd = cmdFormat.parseCommand(cmdString);
+            if (cmd.type == 'W') {
+                for (auto eraseCmd : eraseCmds) {
+                    if (cmd.lba >= eraseCmd.lba
+                        && cmd.lba < eraseCmd.lba + eraseCmd.size) {
+                        fileManager->removeBuffer(eraseCommand(eraseCmd.lba, eraseCmd.size));
+                        int formerSize = cmd.lba - eraseCmd.lba;
+                        if(formerSize > 0) 
+                            fileManager->writeBuffer(eraseCommand(eraseCmd.lba, formerSize));
+                        int latterStart = cmd.lba + 1;
+                        int latterSize = eraseCmd.lba + eraseCmd.size - latterStart;
+                        if(latterSize > 0)
+                            fileManager->writeBuffer(eraseCommand(latterStart, latterSize));
+                    }
+                }
+            }
+        }
+    }
+
+
 
     void flush() {
         PRINT_LOG(("Flush Start"));
-        vector<string> cmdStrings = bufferManager->readBuffer();
+        vector<string> cmdStrings = fileManager->readBuffer();
         for (auto cmdString : cmdStrings) {
-            int lba = bufferManager->getCmdLba(cmdString);
-            string value = "";
-            int size;
-            switch (bufferManager->getCmdType(cmdString)) {
+            Command cmd = cmdFormat.parseCommand(cmdString);
+            switch (cmd.type) {
             case 'R':
-                fileManager->readNand(lba);
+                fileManager->readNand(cmd.lba);
                 break;
             case 'W':
-                value = bufferManager->getCmdValue(cmdString);
-                fileManager->writeNand(lba, value);
+                fileManager->writeNand(cmd.lba, cmd.value);
                 break;
             case 'E':
-                size = bufferManager->getCmdSize(cmdString);
-                for (unsigned int offset = 0; offset < size; offset++) {
-                    fileManager->writeNand((lba + offset), DEFAULT_VALUE);
+                for (unsigned int offset = 0; offset < cmd.size; offset++) {
+                    fileManager->writeNand((cmd.lba + offset), DEFAULT_VALUE);
                 }
                 break;
             default:
                 break;
             }
         }
-        bufferManager->flushBuffer();
+        fileManager->flushBuffer();
         PRINT_LOG(("Flush Finished"));
     }
     FileManager* fileManager;
-    BufferManager* bufferManager;
+    CommandFormat cmdFormat;
+
 private:
     const string DEFAULT_VALUE = "0x00000000";
     const unsigned int MAX_COMMAND_NUM_IN_BUFFER = 10;
