+++ Sample-Test1/test.cpp
@@ -8,7 +8,6 @@
 #include "../SSD_Project/Ssd.cpp"
 #include "../SSD_Project/CommandManager.cpp"
 #include "../SSD_Project/FileManager.cpp"
-#include "../SSD_Project/BufferManager.cpp"
 #include <string>
 #include <iostream>
 
@@ -314,9 +313,8 @@ protected:
 	}
 public:
 	Ssd ssd;
-	FileManager& fileManager = FileManager::getInstance();
-	BufferManager& bufferManager = BufferManager::getInstance();
 
+	FileManager& fileManager = FileManager::getInstance();
 	const string DATA1 = "0x00001004";
 	const string DATA2 = "0x00C0FFEE";
 	const string DEFAULT_WRITE_VALUE = "0x00000000";
@@ -480,7 +478,6 @@ TEST_F(FileManagerFixture, WriteReadTestConsecutiveWrite) {
 	EXPECT_THAT(getResultValue(), StrEq(DATA2));
 }
 
-
 TEST_F(FileManagerFixture, CommandBufferNotExist) {
 	const string cmd = "W 20 0x1234ABCD";
 	EXPECT_THROW(isInCommandBuffer(cmd), FileNotExistException);
@@ -489,7 +486,7 @@ TEST_F(FileManagerFixture, CommandBufferNotExist) {
 TEST_F(FileManagerFixture, WriteCommandBufferFile) {
 	const string cmd = "W 20 0x1234ABCD";
 
-	bufferManager.writeBuffer(cmd);
+	fileManager.writeBuffer(cmd);
 
 	bool expected = true;
 	bool actual = isInCommandBuffer(cmd);
@@ -501,8 +498,8 @@ TEST_F(FileManagerFixture, WriteCommandBufferFileMoreThanOneCommand) {
 	const string cmd1 = "W 20 0x1234ABCD";
 	const string cmd2 = "W 21 0x1234ABCD";
 
-	bufferManager.writeBuffer(cmd1);
-	bufferManager.writeBuffer(cmd2);
+	fileManager.writeBuffer(cmd1);
+	fileManager.writeBuffer(cmd2);
 
 	bool expected = true;
 	bool actual1 = isInCommandBuffer(cmd1);
@@ -517,8 +514,8 @@ TEST_F(FileManagerFixture, WriteCommandBufferFileMoreThanTenCommands) {
 	const string cmd2 = "W 10 0x1234AB02";
 
 	for(int i = 0 ; i<10 ; i++)
-		bufferManager.writeBuffer(cmd1);
-	bufferManager.writeBuffer(cmd2);
+		fileManager.writeBuffer(cmd1);
+	fileManager.writeBuffer(cmd2);
 
 	bool expected1 = false;
 	bool actual1 = isInCommandBuffer(cmd1);
@@ -532,8 +529,8 @@ TEST_F(FileManagerFixture, WriteCommandBufferFileMoreThanTenCommands) {
 
 TEST_F(FileManagerFixture, WriteAndReadCommandBufferFile) {
 	const string cmd = "W 20 0x1234ABCD";
-	bufferManager.writeBuffer(cmd);
-	EXPECT_THAT(bufferManager.readBuffer(), Contains(cmd));
+	fileManager.writeBuffer(cmd);
+	EXPECT_THAT(fileManager.readBuffer(), Contains(cmd));
 }
 
 TEST_F(FileManagerFixture, writeToCommandBuffer) {
@@ -557,7 +554,7 @@ TEST_F(FileManagerFixture, WriteCommandBuferAndReadFromCommandBuffer) {
 	const string cmd = "W " + to_string(line) + " " + writeValue;
 
 	// Act
-	bufferManager.writeBuffer(cmd);
+	fileManager.writeBuffer(cmd);
 	ssd.read(line);
 
 	string expected = writeValue;
@@ -577,8 +574,8 @@ TEST_F(FileManagerFixture, WriteCommandBuferTwiceAndReadFromCommandBuffer) {
 	const string cmd2 = "W " + to_string(line2) + " " + writeValue2;
 
 	// Act
-	bufferManager.writeBuffer(cmd);
-	bufferManager.writeBuffer(cmd2);
+	fileManager.writeBuffer(cmd);
+	fileManager.writeBuffer(cmd2);
 	ssd.read(line);
 
 	string expected = writeValue;
@@ -611,7 +608,7 @@ TEST_F(FileManagerFixture, WriteCommandBufferAndErase) {
 	const string cmd = "W " + to_string(line) + " " + writeValue;
 
 	// Act
-	bufferManager.writeBuffer(cmd);
+	fileManager.writeBuffer(cmd);
 	ssd.erase(line, 1);
 
 	// Assert
@@ -628,8 +625,8 @@ TEST_F(FileManagerFixture, WriteCommandBufferMoreThanOnceAndErase) {
 	const string cmd2 = "W " + to_string(line2) + " " + writeValue2;
 
 	// Act
-	bufferManager.writeBuffer(cmd);
-	bufferManager.writeBuffer(cmd2);
+	fileManager.writeBuffer(cmd);
+	fileManager.writeBuffer(cmd2);
 	ssd.erase(line, 1);
 
 	// Assert
@@ -637,7 +634,7 @@ TEST_F(FileManagerFixture, WriteCommandBufferMoreThanOnceAndErase) {
 	EXPECT_TRUE(isInCommandBuffer(cmd2));
 
 	// Act
-	bufferManager.writeBuffer(cmd);
+	fileManager.writeBuffer(cmd);
 	ssd.erase(line, 2);
 
 	// Assert
@@ -652,7 +649,7 @@ TEST_F(FileManagerFixture, WriteCommandBufferTenTimesAndErase) {
 	const string cmd = "W " + to_string(line) + " " + writeValue;
 	// Act
 	for (int i = 0; i < 10; i++)
-		bufferManager.writeBuffer(cmd);
+		fileManager.writeBuffer(cmd);
 	ssd.erase(line, 1);
 
 	// Assert
@@ -727,7 +724,7 @@ TEST_F(FileManagerFixture, avoidDuplicateWriteToCommandBuffer) {
 	// Act
 	ssd.write(line, writeValue);
 	ssd.write(line, writeValue);
-	vector<string> cmdStrings = bufferManager.readBuffer();
+	vector<string> cmdStrings = fileManager.readBuffer();
 	
 	// Assert
 	EXPECT_EQ(cmdStrings.size(), 1);
@@ -745,7 +742,7 @@ TEST_F(FileManagerFixture, updateWriteToCommandBufferWithNewValue) {
 	// Act
 	ssd.write(line, writeValue);
 	ssd.write(line, writeValue2);
-	vector<string> cmdStrings = bufferManager.readBuffer();
+	vector<string> cmdStrings = fileManager.readBuffer();
 	
 	// Assert
 	EXPECT_EQ(cmdStrings.size(), 1);
@@ -761,7 +758,7 @@ TEST_F(FileManagerFixture, mergeTwoErasesToOneErase) {
 	// Act
 	ssd.erase(line, 1);
 	ssd.erase(line + 1, 1);
-	vector<string> cmdStrings = bufferManager.readBuffer();
+	vector<string> cmdStrings = fileManager.readBuffer();
 
 	// Assert
 	EXPECT_EQ(cmdStrings.size(), 1);
@@ -776,7 +773,7 @@ TEST_F(FileManagerFixture, mergeTwoErasesToOneEraseDifferentOrder) {
 	// Act
 	ssd.erase(line + 1, 1);
 	ssd.erase(line, 1);
-	vector<string> cmdStrings = bufferManager.readBuffer();
+	vector<string> cmdStrings = fileManager.readBuffer();
 
 	// Assert
 	EXPECT_EQ(cmdStrings.size(), 1);
@@ -798,7 +795,7 @@ TEST_F(FileManagerFixture, mergeErasesComplexCase) {
 	const string cmd2 = "E 7 3";
 	const string cmd3 = "E 15 1";
 
-	vector<string> cmdStrings = bufferManager.readBuffer();
+	vector<string> cmdStrings = fileManager.readBuffer();
 
 	// Assert
 	EXPECT_EQ(cmdStrings.size(), 3);
@@ -819,7 +816,7 @@ TEST_F(FileManagerFixture, erasePPTCondition) {
 	const string cmd = "W 20 0xABCDABCD";
 	const string cmd2 = "E 10 5";
 
-	vector<string> cmdStrings = bufferManager.readBuffer();
+	vector<string> cmdStrings = fileManager.readBuffer();
 
 	// Assert
 	EXPECT_EQ(cmdStrings.size(), 2);
@@ -838,7 +835,7 @@ TEST_F(FileManagerFixture, eraseSizeBiggerThan10) {
 
 	const string cmd = "E 8 10";
 
-	vector<string> cmdStrings = bufferManager.readBuffer();
+	vector<string> cmdStrings = fileManager.readBuffer();
 
 	// Assert
 	EXPECT_EQ(cmdStrings.size(), 1);
@@ -860,7 +857,7 @@ TEST_F(FileManagerFixture, optimizeCommandBufferPPT) {
 	const string cmd = "E 10 2";
 
 	// Assert
-	EXPECT_THAT(bufferManager.readBuffer(), Contains(cmd));
+	EXPECT_THAT(fileManager.readBuffer(), Contains(cmd));
 }
 
 TEST_F(FileManagerFixture, optimizeCommandBufferPPT2) {
@@ -872,7 +869,7 @@ TEST_F(FileManagerFixture, optimizeCommandBufferPPT2) {
 	ssd.write(50, "0xABCD1234");
 	const string cmd = "E 40 5";
 	const string cmd2 = "W 50 0xABCD1234";
-	vector<string> cmdStrings = bufferManager.readBuffer();
+	vector<string> cmdStrings = fileManager.readBuffer();
 
 	// Assert
 	EXPECT_EQ(cmdStrings.size(), 2);
@@ -890,7 +887,7 @@ TEST_F(FileManagerFixture, optimizeCommandBufferPPT3) {
 	ssd.erase(12, 3);
 	const string cmd = "W 10 " + writeValue;
 	const string cmd2 = "E 11 4";
-	vector<string> cmdStrings = bufferManager.readBuffer();
+	vector<string> cmdStrings = fileManager.readBuffer();
 	ssd.read(10);
 
 	// Assert
@@ -910,7 +907,7 @@ TEST_F(FileManagerFixture, EraseSameCommandTwice) {
 	ssd.erase(10, 2);
 
 	const string cmd = "E 10 2";
-	vector<string> cmdStrings = bufferManager.readBuffer();
+	vector<string> cmdStrings = fileManager.readBuffer();
 
 	// Assert
 	EXPECT_EQ(cmdStrings.size(), 1);
@@ -925,7 +922,7 @@ TEST_F(FileManagerFixture, WriteSameCommandTwice) {
 	ssd.write(10, writeValue);
 
 	const string cmd = "W 10 " + writeValue;
-	vector<string> cmdStrings = bufferManager.readBuffer();
+	vector<string> cmdStrings = fileManager.readBuffer();
 
 	// Assert
 	EXPECT_EQ(cmdStrings.size(), 1);
